Split ColorscalePicker
