+++ src/components/fields/ColorscalePicker.js
@@ -2,9 +2,9 @@ import ColorscalePicker from '../widgets/ColorscalePicker';
 import Field from './Field';
 import PropTypes from 'prop-types';
 import React, {Component} from 'react';
-import {connectToContainer, adjustColorscale} from 'lib';
+import {connectToContainer} from 'lib';
 
-class UnconnectedColorscalePicker extends Component {
+export class UnconnectedColorscalePicker extends Component {
   constructor(props) {
     super(props);
     this.onUpdate = this.onUpdate.bind(this);
@@ -12,18 +12,6 @@ class UnconnectedColorscalePicker extends Component {
 
   onUpdate(colorscale, colorscaleType) {
     if (Array.isArray(colorscale)) {
-      if (this.context.container.type === 'pie') {
-        const numPieSlices = this.context.graphDiv.calcdata[0].length + 1;
-        const adjustedColorscale = adjustColorscale(
-          colorscale,
-          numPieSlices,
-          colorscaleType,
-          {repeat: true}
-        );
-        this.props.updatePlot(adjustedColorscale);
-        return;
-      }
-
       this.props.updatePlot(
         colorscale.map((c, i) => {
           let step = i / (colorscale.length - 1);
@@ -67,45 +55,4 @@ UnconnectedColorscalePicker.contextTypes = {
   graphDiv: PropTypes.object,
 };
 
-export default connectToContainer(UnconnectedColorscalePicker, {
-  modifyPlotProps: (props, context, plotProps) => {
-    if (
-      props.attr === 'marker.color' &&
-      context.fullData
-        .filter(t => context.traceIndexes.includes(t.index))
-        .every(t => t.marker && t.marker.color) &&
-      (plotProps.fullValue && typeof plotProps.fullValue === 'string')
-    ) {
-      plotProps.fullValue =
-        context.fullData &&
-        context.fullData
-          .filter(t => context.traceIndexes.includes(t.index))
-          .map(t => [0, t.marker.color]);
-    }
-
-    if (
-      context &&
-      context.container &&
-      context.graphDiv &&
-      (!plotProps.fullValue ||
-        (Array.isArray(plotProps.fullValue) && !plotProps.fullValue.length)) &&
-      context.container.type === 'pie' &&
-      context.graphDiv.calcdata
-    ) {
-      plotProps.fullValue = context.graphDiv.calcdata[0].map(d => [0, d.color]);
-    }
-
-    if (
-      props.attr === 'marker.colors' &&
-      plotProps.fullValue &&
-      Array.isArray(plotProps.fullValue) &&
-      !plotProps.fullValue.every(el => Array.isArray(el))
-    ) {
-      plotProps.fullValue = plotProps.fullValue.map(c => [0, c]);
-    }
-
-    if (context.container.type === 'pie' && context.traceIndexes.length > 1) {
-      plotProps.isVisible = false;
-    }
-  },
-});
+export default connectToContainer(UnconnectedColorscalePicker);
