+++ src/components/index.js
@@ -12,6 +12,7 @@ import {
   AxesSelector,
   ColorPicker,
   ColorscalePicker,
+  PieColorscalePicker,
   ColorwayPicker,
   ContourNumeric,
   DataSelector,
@@ -110,6 +111,7 @@ export {
   ColorPicker,
   ColorscalePicker,
   ColorwayPicker,
+  PieColorscalePicker,
   ContourNumeric,
   DataSelector,
   Dropdown,
