+++ src/index.js
@@ -38,6 +38,7 @@ import {
   AxesSelector,
   Button,
   ColorPicker,
+  PieColorscalePicker,
   ColorscalePicker,
   ContourNumeric,
   ErrorBars,
@@ -107,6 +108,7 @@ export {
   Button,
   ColorPicker,
   ColorscalePicker,
+  PieColorscalePicker,
   ContourNumeric,
   ErrorBars,
   DataSelector,
